refactor: move the api date helper to utils
